bugfix: 添加设备管理, 显示删除习题按钮
        复用5000组播端口
        学生端监听5987(接收教师端命令)
        教师端监听5986(接收学生状态)
        收发PC和投屏状态用组播5000

      添加了第1\2章生物的内容, 修正mq地址
      * 修复下划线问题
      * 修复图片排版问题
      * 互换了拍照和重拍按钮位置
      * 拍照加入新题，自动跳转到最后一题
      * 修改搜题失败时的提示
      * 解决多次删除题目会崩溃问题
      近期更新:
      * 可以从登录界面切换测试版
      * 增加AI切换角色
      * 搜题失败时增加跳转文字搜索接口
      * 增加AI对话角色选择按钮
      - 待完成: 新消息小红点

+++ app/src/main/java/com/cosinetech/imates/activities/QuestionSolveActivity.java
@@ -1,5 +1,6 @@
 package com.cosinetech.imates.activities;
 
+import android.app.AlertDialog;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.KeyEvent;
@@ -12,6 +13,7 @@ import android.widget.Toast;
 import androidx.activity.EdgeToEdge;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.content.res.AppCompatResources;
 import androidx.lifecycle.ViewModelProvider;
 import androidx.lifecycle.ViewModelStoreOwner;
 import androidx.recyclerview.widget.LinearLayoutManager;
@@ -32,6 +34,7 @@ import com.cosinetech.imates.models.FindSimilarQuestionRequest;
 import com.cosinetech.imates.models.Subject;
 import com.cosinetech.imates.models.UserInfoViewModel;
 import com.cosinetech.imates.mq.MessagingManager;
+import com.cosinetech.imates.screencasting.ScreenCastingManager;
 import com.cosinetech.imates.util.AppUtils;
 import com.cosinetech.imates.util.WindowUtils;
 import com.cosinetech.imates.views.ChatAiView;
@@ -41,6 +44,8 @@ import com.cosinetech.imates.webservice.AiChatMessageRequest;
 import com.cosinetech.imates.webservice.ApiGateWayService;
 import com.cosinetech.imates.webservice.ApiUrl;
 import com.cosinetech.imates.webservice.Question;
+
+import org.loka.screensharekit.ScreenShareKit;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -172,31 +177,40 @@ public class QuestionSolveActivity extends AppCompatActivity implements Messagin
                 if(position < 0 || position >= mQuestions.size()) {
                     return;
                 }
-                String url;
-                Question q = mQuestions.get(position);
-                if(subject == Subject.SUBJECT_BIOLOGY) {
-                    url = ApiUrl.URL_DELETE_EXERCISE_BASE + "/" + q.id + "/biology";
-                } else if(subject == Subject.SUBJECT_MATH) {
-                    url = ApiUrl.URL_DELETE_EXERCISE_BASE + "/" + q.id + "/math";
-                } else {
-                    return;
+                new AlertDialog.Builder(QuestionSolveActivity.this)
+                        .setTitle("提示")
+                        .setMessage("确认删除习题吗?")
+                        .setPositiveButton("确认", (dialog, which) -> {
+                            String url;
+                            Question q = mQuestions.get(position);
+                            if(subject == Subject.SUBJECT_BIOLOGY) {
+                                url = ApiUrl.URL_DELETE_EXERCISE_BASE + "/" + q.id + "/biology";
+                            } else if(subject == Subject.SUBJECT_MATH) {
+                                url = ApiUrl.URL_DELETE_EXERCISE_BASE + "/" + q.id + "/math";
+                            } else {
+                                return;
+                            }
+                            ApiGateWayService.deleteExercise(url, userInfoViewModel.token.getValue(), new ApiGateWayService.ExerciseDeleteLister() {
+                                @Override
+                                public void onDeleteSuccess() {
+                                    runOnUiThread(() -> {
+                                        Question q = mQuestions.remove(position);
+                                        adapterQuestionList.notifyItemRemoved(position);
+                                    });
+
+                                }
+
+                                @Override
+                                public void onDeleteFailed(String msg) {
+                                    runOnUiThread(() -> Toast.makeText(QuestionSolveActivity.this, "删除失败, 稍后重试", Toast.LENGTH_SHORT).show());
+                                }
+                            });
+                        })
+                        .setNegativeButton("取消", (dialog, which) -> {
+                        })
+                        .create()
+                        .show();
                 }
-                ApiGateWayService.deleteExercise(url, userInfoViewModel.token.getValue(), new ApiGateWayService.ExerciseDeleteLister() {
-                    @Override
-                    public void onDeleteSuccess() {
-                        runOnUiThread(() -> {
-                            Question q = mQuestions.remove(position);
-                            adapterQuestionList.notifyItemRemoved(position);
-                        });
-
-                    }
-
-                    @Override
-                    public void onDeleteFailed(String msg) {
-                        runOnUiThread(() -> Toast.makeText(QuestionSolveActivity.this, "删除失败, 稍后重试", Toast.LENGTH_SHORT).show());
-                    }
-                });
-            }
 
             @Override
             public void onExerciseToTop(int position) {
